Move subview set-up to another file.
